requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-socketio[asyncio]>=5.11.0
aiohttp>=3.9.0
websockets>=12.0
anthropic>=0.28.0
openai>=1.30.0
ollama>=0.2.0
zeroconf>=0.132.0
bleak>=0.21.0
pyyaml>=6.0.1
jinja2>=3.1.4
python-multipart>=0.0.9
aiosqlite>=0.20.0
